Allow permutation to return just one permutation.

+++ theano/tensor/tests/test_opt.py
@@ -3889,7 +3889,8 @@ class T_local_reduce(unittest.TestCase):
             assert isinstance(topo[-1].op, T.Elemwise), out
 
         # Test different axis for the join and the reduction
-        A = theano.shared(numpy.array([1, 2, 3, 4, 5]))
+        # We must force the dtype, of otherwise, this tests will fail in 32 bit system
+        A = theano.shared(numpy.array([1, 2, 3, 4, 5], dtype='int64'))
 
         f = theano.function([], T.sum(T.stack(A, A), axis=0), mode=self.mode)
         assert numpy.allclose(f(), [2, 4, 6, 8, 10])
